Ai/image_captioning/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
